+++ lib/input/percentageInput.d.ts
@@ -1,6 +1,6 @@
 import m, { CVnode } from "mithril";
 import { IPropWidget } from "../interface/widget";
-import { ValidationBase } from "../validationBase";
-export declare class PercentageInput extends ValidationBase<IPropWidget> {
+import { BaseWidget } from "../baseWidget";
+export declare class PercentageInput extends BaseWidget<IPropWidget> {
     view({ attrs }: CVnode<IPropWidget>): m.Vnode<import("../interface/widget").IPropLayoutWidget<import("../interface/widget").IField>, unknown>;
 }