+++ lib/input/radio.d.ts
@@ -1,7 +1,8 @@
-import m, { ClassComponent, CVnode } from "mithril";
+import m, { CVnode } from "mithril";
 import { IPropWidget, IRadioField } from "../interface/widget";
+import { BaseWidget } from "../baseWidget";
 type TRadioWidget = IPropWidget<IRadioField>;
-export declare class RadioInput implements ClassComponent<TRadioWidget> {
+export declare class RadioInput extends BaseWidget<TRadioWidget> {
     view({ attrs }: CVnode<TRadioWidget>): m.Vnode<any, any>;
 }
 export {};