+++ lib/input/timeInput.d.ts
@@ -1,7 +1,7 @@
 import m, { CVnode } from "mithril";
 import { IPropWidget } from "../interface/widget";
-import { ValidationBase } from "../validationBase";
-export declare class TimeInput extends ValidationBase<IPropWidget> {
+import { BaseWidget } from "../baseWidget";
+export declare class TimeInput extends BaseWidget<IPropWidget> {
     private showPicker;
     private focus;
     private readonly hour;